Sort series with decimal sequence numbers correctly
Series sequences are free-form strings on the server and are often
fractional ("1.1", "1.2"). The domain model, database column and
mapping layer all treated the sequence as an Int, so any decimal
sequence failed to parse, fell back to Int.MAX_VALUE and sorted to the
end of the series.

Make the sequence a Double end-to-end: domain model, REAL column
(migration 9 rebuilds the media table since SQLite cannot alter a
column type), JSON list adapter, mapping, fallbacks and the session
payload formatter. Sequences with a numeric prefix ("2a", "4-5") use
the leading number.

Fixes #956

+++ features/home/impl/src/commonMain/kotlin/app/campfire/home/store/home/HomeSourceOfTruthFactory.kt
@@ -256,7 +256,7 @@ class HomeSourceOfTruthFactory(
       // Insert the series books
       series.books?.forEachIndexed { index, book ->
         val libraryItem = book.asDbModel()
-        val media = book.media.asDbModel(book.id, fallbackSeriesSequence = index + 1)
+        val media = book.media.asDbModel(book.id, fallbackSeriesSequence = (index + 1).toDouble())
 
         // If these items exist, lets not overwrite their metadata
         db.libraryItemsQueries.insertOrIgnore(libraryItem)
